copyteximage: Test the GL_STENCIL_INDEX8 format
The copyteximage test now tests the GL_STENCIL_INDEX8 format along
with all of the other formats. This requires a shader because the
stencil index values can only be accessed from an integer sampler. The
shader is now set up outside of the per-target switch statement
because it is now potentially needed by all targets.