point-vertex-id: Additionally test with gl_InstanceID
The point-vertex-id test was originally created to test gl_VertexID in
combination with glPolygonMode(GL_POINT) because of a bug with the
i965 driver in Mesa. The same bug exists with gl_InstanceID and the
code for BDW has some extra complications in this case so I think it's
worth extending the test to check that as well. The test now takes a
command line argument to specify which builtin to test. It can also
take both arguments to generate a shader which relies on both.