point-vertex-id: Test using an instanced attribute
Adds an option to use an instanced attribute so that it can be tested
in combination with glPolygonMode. This demonstrates a bug on the Mesa
i965 driver on BDW which was reported as bug #91292. The ‘divisor’
argument can be added in addition to any of the other arguments so
that all of the combinations can be test.