Test that interpolateAtSample returns different results
Adds a new test for interpolateAtSample which captures the results
from the function with each different sample value and ensures that
they are all different. This can be done either with a constant or
non-constant value depending on whether a command line argument is
given.

The other tests for interpolateAtSample always use 0 as the sample ID
so it's not very thorough. Using non-const sampler IDs isn't
implemented in the i965 driver of Mesa but the other tests pass anyway
on a release build because it is hardcoded to use sample 0.

Reviewed-by: Dave Airlie <airlied@redhat.com>