Test interpolateAtSample with dynamically nonuniform values
Adds a test which is similar to interpolateAtSample-different except
that it draws a triangle which covers more than one fragment and makes
sure to use a different sample ID for each fragment so that it won't
be dynamically uniform. The GLSL spec doesn't mention that the sample
ID has to be dynamically uniform which presumably means that it is
allowed. I think this is worth testing because it is not
straightforward to implement at least on Intel hardware.