Demonstrate a bug with fast clears on MS FBOs with missing comps
There is currently a bug with the i965 driver in Mesa with the fast
clear optimisation when there are missing components in the
framebuffer. If the texture that was rendered to is later sampled it
will use the fast clear color without overriding the missing
components to 0 or 1. This patch adds a test to demonstrate that.