shaders: Test using a bound program after unsuccessful relink
If a bound program is relinked unsuccessfully it is supposed to keep
the executable and the program state such as the uniforms. This adds a
test for that which updates a uniform, does a failed relink and
verifies that a render succeeds with the new uniform value. This is
currently causing a use-after-free error in Mesa with a failure rate
of about 3 out of 10.